Update assignment7b.html
Assignment description: 

Assignment 7_B Sample Page: I posted a word doc in the resources folder. Duplicate that doc in html. Use any image you want. Add a background color and designate a color for the text. Upload the new page and the image to GitHub. Add a link on your index page to this new page. Due October 8 (5 points)